backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
groq==0.9.0
pinecone-client==3.2.2
sentence-transformers==3.0.0
pymupdf==1.24.3
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.7.1
